# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Report Generator API
  version: 1.0.0
  description: API for developer report generation. Syncs data from GitLab and Bitrix24, matches commits to tasks, generates narratives via LLM, exports to .docx.
servers:
  - url: http://localhost:8000/api
    description: Local development
paths:
...
  /settings:
    get:
      tags: [Settings]
      operationId: getSettings
      summary: Get current application settings
      responses:
        '200':
          description: Current settings
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Settings'
        '500':
          $ref: '#/components/responses/InternalError'

    put:
      tags: [Settings]
      operationId: updateSettings
      summary: Update application settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SettingsInput'
      responses:
        '200':
          description: Settings updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Updated
                  data:
                    $ref: '#/components/schemas/Settings'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalError'

# ── Components ────────────────────────────────────────────────────────────
components:
  schemas:
...
    Settings:
      type: object
      properties:
        gitlab_token_set:
          type: boolean
          description: Whether GitLab token is configured (actual value is never exposed)
        gitlab_username:
          type: string
          nullable: true
        bitrix24_api_key_set:
          type: boolean
          description: Whether Bitrix24 API key is configured
        bitrix24_user_id:
          type: string
          nullable: true
        llm_provider:
          type: string
          enum: [claude, openai]
          nullable: true
        llm_api_key_set:
          type: boolean
        llm_system_prompt:
          type: string
          nullable: true
        developer_name:
          type: string
          nullable: true
        developer_position:
          type: string
          nullable: true
        sync_schedule_time:
          type: string
          nullable: true
          example: '02:00'
    SettingsInput:
      type: object
      properties:
        gitlab_token:
          type: string
          nullable: true
        gitlab_username:
          type: string
          nullable: true
        bitrix24_api_key:
          type: string
          nullable: true
        bitrix24_user_id:
          type: string
          nullable: true
        llm_provider:
          type: string
          enum: [claude, openai]
          nullable: true
        llm_api_key:
          type: string
          nullable: true
        llm_system_prompt:
          type: string
          nullable: true
        developer_name:
          type: string
          nullable: true
        developer_position:
          type: string
          nullable: true
        sync_schedule_time:
          type: string
          nullable: true
    ErrorResponse:
      type: object
      required: [code, message]
      properties:
        code:
          type: integer
          example: 422
        message:
          type: string
          example: Validation failed
        errors:
          type: object
          additionalProperties:
            type: array
            items:
              type: string
          example:
            date_from: ['The date_from field is required.']
...
  responses:
...
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            code: 500
            message: Internal server error